Fix PlantUML link for context diagram and container diagram (#4)
This specified `test-plantuml` which is a branch that only existed
for the PR which meant the `src` we were provided doesn't exist
anymore.

Since we are mainly concerned with it working on `main`, it's just been
updated to hardcode that in. However, this does mean it won't work
for other branches.